refactor: フィルタリング処理を整理
