+++ public/iframe.html
@@ -195,19 +195,19 @@
       <!-- 絵文字タブ -->
       <div class="tab-pane fade pb-2 px-2" id="emojis">
         <div class="d-flex overflow-x-auto btn-container gap-1" role="group">
-          <button type="button" class="btn text-white-50 flex-shrink-0 p-1 emoji-item" data-emoji="😀" title="スマイル">
+          <button type="button" class="btn text-white-50 flex-shrink-0 p-1 emoji-item" data-emoji="😀" title="スマイル" data-text="スマイル">
             <span class="text-custom d-block">😀</span>
           </button>
-          <button type="button" class="btn text-white-50 flex-shrink-0 p-1 emoji-item" data-emoji="😂" title="笑い">
+          <button type="button" class="btn text-white-50 flex-shrink-0 p-1 emoji-item" data-emoji="😂" title="笑い" data-text="笑い">
             <span class="text-custom d-block">😂</span>
           </button>
-          <button type="button" class="btn text-white-50 flex-shrink-0 p-1 emoji-item" data-emoji="😍" title="ハート目">
+          <button type="button" class="btn text-white-50 flex-shrink-0 p-1 emoji-item" data-emoji="😍" title="ハート目" data-text="ハート目">
             <span class="text-custom d-block">😍</span>
           </button>
-          <button type="button" class="btn text-white-50 flex-shrink-0 p-1 emoji-item" data-emoji="👍" title="いいね">
+          <button type="button" class="btn text-white-50 flex-shrink-0 p-1 emoji-item" data-emoji="👍" title="いいね" data-text="いいね">
             <span class="text-custom d-block">👍</span>
           </button>
-          <button type="button" class="btn text-white-50 flex-shrink-0 p-1 emoji-item" data-emoji="🙏" title="お願い">
+          <button type="button" class="btn text-white-50 flex-shrink-0 p-1 emoji-item" data-emoji="🙏" title="お願い" data-text="お願い">
             <span class="text-custom d-block">🙏</span>
           </button>
         </div>
