+++ src/content/emojis.ts
@@ -1,3 +1,5 @@
+import { performFiltering, sortByMatchPosition } from './utils/filtering';
+
 /** ツールアイテムのクリックイベントを設定 */
 export function setupEmojiItemListeners(iframeDoc: Document): void {
   const emojiItems = iframeDoc.querySelectorAll<HTMLElement>('.emoji-item');
@@ -18,4 +20,14 @@ export function setupEmojiItemListeners(iframeDoc: Document): void {
       }
     });
   });
+}
+
+/** 絵文字アイテムを検索してフィルタリング */
+export function filterEmojiItems(iframeDoc: Document, query: string): void {
+  const matchedItems = performFiltering(iframeDoc, query, '.emoji-item');
+
+  // クエリが空でない場合、マッチ位置でソート
+  if (query !== '' && matchedItems.length > 0) {
+    sortByMatchPosition(matchedItems);
+  }
 }
